+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Labels in the navigation can now be translated. If the navigation-specific translation key isn't set, we'll use the title translation for the repositorys index page.
+- The button to add a new model now falls back to the translation from  common.new when the button-specific translation doesn't exist before falling back to "New %{model}".
 
 ### Fixed